refactor: finish DeviceClient architecture cleanup
Move device-domain mixins out of coordinator/ into core/:
- core/io_mixin.py: _ModbusIOMixin (Modbus read-path protocol, no HA deps)
- core/capabilities_mixin.py: _CoordinatorCapabilitiesMixin (derived metrics, no HA deps)

Both coordinator/ files become thin re-export shims so existing imports
continue to work without changes, including the test_clock_sync.py import.

core/client.py now imports these mixins from the local core/ package instead
of reaching into coordinator/, making the dependency direction cleaner:
  core/ <- coordinator/ (one-way), not coordinator/ <- core/.

Remove the dynamic coordinator import from DeviceClient connection code:
- client_connection.py._try_direct_client_connect previously did a runtime
  import of coordinator.coordinator to read its AsyncModbusTcpClient name so
  tests could patch it there; the indirection is eliminated by exposing
  AsyncModbusTcpClient as a plain module-level name in client_connection.py
  that mock.patch can target directly.
- coordinator.py drops the now-unused AsyncModbusTcpClient import and the
  _ORIGINAL_ASYNC_MODBUS_TCP_CLIENT alias.
- test_coordinator.py patch target updated:
    coordinator.coordinator.AsyncModbusTcpClient
    -> core.client_connection.AsyncModbusTcpClient
