+++ custom_components/thessla_green_modbus/core/capabilities_mixin.py
@@ -1,0 +1,276 @@
+"""Capabilities and derived metrics mixin for device-domain clients.
+
+Moved from coordinator/capabilities.py so that ThesslaGreenDeviceClient can
+inherit it without creating a coordinator/ → core/ dependency.  The coordinator
+continues to inherit it via coordinator/capabilities.py (which re-exports from
+here).
+"""
+
+from __future__ import annotations
+
+import logging
+from collections.abc import Mapping
+from datetime import UTC, datetime
+from types import MappingProxyType
+from typing import Any, ClassVar
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def _utcnow() -> datetime:
+    """Return timezone-aware UTC now."""
+    return datetime.now(UTC)
+
+
+def _clamp_percentage(value: float) -> float:
+    """Clamp a percentage value into the [0, 100] range."""
+    return max(0.0, min(100.0, value))
+
+
+def _flow_balance_status(balance: float) -> str:
+    """Return flow-balance status label for a computed balance value."""
+    if abs(balance) < 10:
+        return "balanced"
+    return "supply_dominant" if balance > 0 else "exhaust_dominant"
+
+
+def _coerce_bypass_open(raw_bypass: Any) -> bool:
+    """Return whether bypass mode indicates an open bypass damper."""
+    return raw_bypass in (1, 2)
+
+
+def _normalise_capability_flag(raw_value: Any) -> bool:
+    """Normalise capability-like values to a strict boolean."""
+    return bool(raw_value) if raw_value is not None else False
+
+
+class _CoordinatorCapabilitiesMixin:
+    """Capability and derived-value logic shared by DeviceClient and coordinator."""
+
+    device_info: dict[str, Any]
+    _last_power_timestamp: datetime | None
+    _total_energy: float
+
+    _STANDBY_POWER_W: float = 10.0
+    _MODEL_POWER_DATA: ClassVar[Mapping[int, tuple[float, float]]] = MappingProxyType(
+        {
+            300: (105.0, 1150.0),
+            400: (170.0, 1500.0),
+            420: (94.0, 1449.0),
+            500: (255.0, 1850.0),
+            550: (345.0, 1950.0),
+        }
+    )
+    _MODEL_FLOW_TOLERANCE = 15
+
+    def _apply_capability_result(
+        self, data: dict[str, Any], key: str, value: Any, capability_flag: Any
+    ) -> bool:
+        """Apply a derived value only when a capability flag is enabled."""
+        if not _normalise_capability_flag(capability_flag):
+            return False
+        data[key] = value
+        return True
+
+    def _calculate_heat_recovery_efficiency(self, data: dict[str, Any]) -> float | None:
+        """Calculate heat recovery efficiency percentage if inputs allow it."""
+        bypass_open = _coerce_bypass_open(data.get("bypass_mode"))
+        if bypass_open:
+            return None
+        if not all(
+            k in data for k in ("outside_temperature", "supply_temperature", "exhaust_temperature")
+        ):
+            return None
+        outside = float(data["outside_temperature"])
+        supply = float(data["supply_temperature"])
+        exhaust = float(data["exhaust_temperature"])
+        if exhaust - outside < 5.0:
+            return None
+        q_supply = data.get("supply_flow_rate")
+        q_exhaust = data.get("exhaust_flow_rate")
+        if q_supply is not None and q_exhaust is not None:
+            q_s = float(q_supply)
+            q_e = float(q_exhaust)
+            q_min = min(q_s, q_e)
+            raw = (
+                (q_s * (supply - outside)) / (q_min * (exhaust - outside))
+                if q_min > 0
+                else (supply - outside) / (exhaust - outside)
+            )
+        else:
+            raw = (supply - outside) / (exhaust - outside)
+        return _clamp_percentage(round(raw * 100, 1))
+
+    def _apply_post_process_derived_values(self, data: dict[str, Any]) -> None:
+        """Apply grouped derived metrics to mutable coordinator data."""
+        try:
+            efficiency = self._calculate_heat_recovery_efficiency(data)
+            if self._apply_capability_result(
+                data, "calculated_efficiency", efficiency, efficiency is not None
+            ):
+                data["heat_recovery_efficiency"] = data["calculated_efficiency"]
+        except (ZeroDivisionError, TypeError, ValueError) as exc:
+            _LOGGER.debug("Could not calculate efficiency: %s", exc)
+
+        if all(
+            k in data for k in ("supply_flow_rate", "outside_temperature", "supply_temperature")
+        ):
+            try:
+                flow = float(data["supply_flow_rate"])
+                delta_t = float(data["supply_temperature"]) - float(data["outside_temperature"])
+                data["heat_recovery_power"] = round(max(0.0, 0.335 * flow * delta_t), 1)
+            except (TypeError, ValueError) as exc:
+                _LOGGER.debug("Could not calculate heat recovery power: %s", exc)
+
+        if "supply_flow_rate" in data and "exhaust_flow_rate" in data:
+            try:
+                balance = float(data["supply_flow_rate"]) - float(data["exhaust_flow_rate"])
+                data["flow_balance"] = balance
+                data["flow_balance_status"] = _flow_balance_status(balance)
+            except (TypeError, ValueError) as exc:
+                _LOGGER.debug("Could not calculate flow balance: %s", exc)
+
+    def _decode_device_clock(self, data: dict[str, Any]) -> str | None:
+        """Decode device clock from packed BCD date/time registers."""
+        raw_yymm = data.get("date_time")
+        raw_ddtt = data.get("date_time_ddtt")
+        raw_ggmm = data.get("date_time_ggmm")
+        raw_sscc = data.get("date_time_sscc")
+        if any(v is None for v in (raw_yymm, raw_ddtt, raw_ggmm, raw_sscc)):
+            return None
+
+        def _bcd(b: int) -> int:
+            return ((b >> 4) & 0xF) * 10 + (b & 0xF)
+
+        yy = _bcd((raw_yymm >> 8) & 0xFF)
+        mm = _bcd(raw_yymm & 0xFF)
+        dd = _bcd((raw_ddtt >> 8) & 0xFF)
+        hh = _bcd((raw_ggmm >> 8) & 0xFF)
+        mi = _bcd(raw_ggmm & 0xFF)
+        ss = _bcd((raw_sscc >> 8) & 0xFF)
+        year = 2000 + yy
+        if 1 <= mm <= 12 and 1 <= dd <= 31 and hh <= 23 and mi <= 59 and ss <= 59:
+            return f"{year:04d}-{mm:02d}-{dd:02d}T{hh:02d}:{mi:02d}:{ss:02d}"
+        return None
+
+    def _lookup_model_power(self, nominal_flow: float) -> tuple[float, float] | None:
+        """Return (fan_total_max_W, heater_max_W) for the closest known model.
+
+        Returns ``None`` when no entry is within ``_MODEL_FLOW_TOLERANCE`` m³/h.
+        """
+        best: tuple[int, tuple[float, float]] | None = None
+        for flow_key, specs in self._MODEL_POWER_DATA.items():
+            diff = abs(nominal_flow - flow_key)
+            if diff <= self._MODEL_FLOW_TOLERANCE:
+                if best is None or diff < abs(nominal_flow - best[0]):
+                    best = (flow_key, specs)
+        return best[1] if best is not None else None
+
+    def calculate_power_consumption(self, data: dict[str, Any]) -> float | None:
+        """Calculate electrical power consumption.
+
+        When the device's nominal airflow is recognised the calculation uses
+        the fan affinity law applied to the *measured* supply and exhaust flow
+        rates together with model-specific max-power values taken from the
+        official datasheet / nameplate.  This gives ±10-15 % accuracy.
+
+        For an unknown model the method falls back to a cubic estimate based
+        on the DAC output voltages (previous behaviour, ±40-50 % accuracy).
+        """
+        nominal_raw = data.get("nominal_supply_air_flow")
+        supply_flow = data.get("supply_flow_rate")
+        exhaust_flow = data.get("exhaust_flow_rate")
+
+        if nominal_raw is not None and supply_flow is not None and exhaust_flow is not None:
+            try:
+                nominal = float(nominal_raw)
+                q_s = max(0.0, float(supply_flow))
+                q_e = max(0.0, float(exhaust_flow))
+                specs = self._lookup_model_power(nominal)
+                if specs is not None and nominal > 0:
+                    fan_total_max, heater_max = specs
+                    # Fan affinity law: P ∝ Q³, split equally between both fans.
+                    fan_per = fan_total_max / 2.0
+                    p_fans = fan_per * (q_s / nominal) ** 3 + fan_per * (q_e / nominal) ** 3
+
+                    # Heater: PWM-controlled resistive element → linear with DAC voltage.
+                    dac_h = float(data.get("dac_heater", 0) or 0)
+                    dac_h = max(0.0, min(10.0, dac_h))
+                    p_heater = heater_max * (dac_h / 10.0)
+
+                    return round(p_fans + p_heater + self._STANDBY_POWER_W, 1)
+            except (TypeError, ValueError) as exc:
+                _LOGGER.debug("Power calculation via flow/DAC unavailable: %s", exc)
+
+        # Fallback: DAC-voltage cubic estimate (model unknown or flow unavailable).
+        try:
+            v_s = float(data["dac_supply"])
+            v_e = float(data["dac_exhaust"])
+        except (KeyError, TypeError, ValueError):
+            return None
+
+        def _dac_power(v: float, p_max: float) -> float:
+            v = max(0.0, min(10.0, v))
+            return (v / 10) ** 3 * p_max
+
+        power = _dac_power(v_s, 80.0) + _dac_power(v_e, 80.0)
+        dac_h = float(data.get("dac_heater", 0) or 0)
+        if dac_h:
+            power += _dac_power(dac_h, 2000.0)
+        dac_c = float(data.get("dac_cooler", 0) or 0)
+        if dac_c:
+            power += _dac_power(dac_c, 1000.0)
+        return round(power, 1)
+
+    def _post_process_data(self, data: dict[str, Any]) -> dict[str, Any]:
+        """Post-process data to calculate derived values."""
+        self._apply_serial_number_state(data)
+        self._apply_post_process_derived_values(data)
+        self._apply_power_and_energy_estimates(data)
+        self._apply_device_clock(data)
+        return data
+
+    def _apply_serial_number_state(self, data: dict[str, Any]) -> None:
+        """Expose known device serial number in coordinator data."""
+        # Expose the full serial number (assembled from 6 registers by the scanner)
+        # as a sensor value so the serial_number entity has a meaningful state.
+        device_serial = (self.device_info or {}).get("serial_number")
+        if device_serial and device_serial != "Unknown":
+            data["serial_number"] = device_serial
+
+    def _apply_power_and_energy_estimates(self, data: dict[str, Any]) -> None:
+        """Apply power estimate and accumulate total energy."""
+        power = self.calculate_power_consumption(data)
+        if power is None:
+            return
+        data["estimated_power"] = power
+        data["electrical_power"] = power
+        now = _utcnow()
+        last_ts = self._last_power_timestamp
+        if not isinstance(last_ts, datetime):
+            elapsed = 0.0
+        else:
+            if (
+                getattr(now, "tzinfo", None) is not None
+                and getattr(last_ts, "tzinfo", None) is None
+            ):
+                last_ts = last_ts.replace(tzinfo=UTC)
+            elif (
+                getattr(now, "tzinfo", None) is None
+                and getattr(last_ts, "tzinfo", None) is not None
+            ):
+                now = now.replace(tzinfo=UTC)
+            elapsed = (now - last_ts).total_seconds()
+        self._total_energy += power * elapsed / 3600000.0
+        data["total_energy"] = self._total_energy
+        self._last_power_timestamp = now
+
+    def _apply_device_clock(self, data: dict[str, Any]) -> None:
+        """Decode and store device clock when valid."""
+        # Decode device clock from BCD registers 0-3
+        try:
+            device_clock = self._decode_device_clock(data)
+            if device_clock is not None:
+                data["device_clock"] = device_clock
+        except (TypeError, ValueError, AttributeError) as exc:
+            _LOGGER.debug("Failed to decode device clock: %s", exc)
